renamed app.py to routes.py & fixed login manager
